LPS-54987 Move package
