+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-form-renderer-web/bnd.bnd
@@ -1,0 +1,8 @@
+Bundle-Name: Liferay Dynamic Data Mapping Form Renderer Web
+Bundle-SymbolicName: com.liferay.dynamic.data.mapping.form.renderer.web
+Bundle-Version: 1.0.0
+Export-Package: com.liferay.dynamic.data.mapping.form.renderer.web
+Include-Resource:\
+	classes,\
+	META-INF/resources=src/META-INF/resources
+Web-ContextPath: /ddm-form-renderer-web
