+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-form-renderer-web/src/META-INF/resources/js/field_types.js
