--- modules/apps/dynamic-data-mapping/dynamic-data-mapping-form-renderer/bnd.bnd
@@ -1,8 +1,0 @@
-Bundle-Name: Liferay Dynamic Data Mapping Form Renderer
-Bundle-SymbolicName: com.liferay.dynamic.data.mapping.form.renderer
-Bundle-Version: 1.0.0
-Export-Package: com.liferay.dynamic.data.mapping.form.renderer
-Include-Resource:\
-	classes,\
-	META-INF=src/META-INF
-Web-ContextPath: /ddm-form-renderer
