--- modules/apps/dynamic-data-mapping/dynamic-data-mapping-form-renderer/src/com/liferay/dynamic/data/mapping/form/renderer/DDMFormRendererConstants.java
@@ -1,37 +1,0 @@
-/**
- * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
- *
- * This library is free software; you can redistribute it and/or modify it under
- * the terms of the GNU Lesser General Public License as published by the Free
- * Software Foundation; either version 2.1 of the License, or (at your option)
- * any later version.
- *
- * This library is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
- * details.
- */
-
-package com.liferay.dynamic.data.mapping.form.renderer;
-
-import com.liferay.portal.kernel.util.StringPool;
-
-/**
- * @author Marcellus Tavares
- */
-public class DDMFormRendererConstants {
-
-	public static final String DDM_FORM_FIELD_LANGUAGE_ID_SEPARATOR =
-		StringPool.DOUBLE_DOLLAR;
-
-	public static final String DDM_FORM_FIELD_NAME_PREFIX = "ddm$$";
-
-	public static final String DDM_FORM_FIELD_PARTS_SEPARATOR =
-		StringPool.DOLLAR;
-
-	public static final String DDM_FORM_FIELDS_SEPARATOR = StringPool.POUND;
-
-	private DDMFormRendererConstants() {
-	}
-
-}
